refactor(api): move OrganizationInvitation model
to src
