+++ api/db/database-builder/factory/build-organization-invitation.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { OrganizationInvitation } from '../../../lib/domain/models/OrganizationInvitation.js';
+import { OrganizationInvitation } from '../../../src/team/domain/models/OrganizationInvitation.js';
 import { databaseBuffer } from '../database-buffer.js';
 import { buildOrganization } from './build-organization.js';
