+++ api/lib/domain/models/index.js
@@ -42,6 +42,7 @@ import { Challenge } from '../../../src/shared/domain/models/Challenge.js';
 import { Competence } from '../../../src/shared/domain/models/Competence.js';
 import { Examiner } from '../../../src/shared/domain/models/Examiner.js';
 import { User } from '../../../src/shared/domain/models/User.js';
+import { OrganizationInvitation } from '../../../src/team/domain/models/OrganizationInvitation.js';
 import { CampaignParticipant } from './../../../src/prescription/campaign-participation/domain/models/CampaignParticipant.js';
 import { CampaignParticipation } from './../../../src/prescription/campaign-participation/domain/models/CampaignParticipation.js';
 import { AccountRecoveryDemand } from './AccountRecoveryDemand.js';
@@ -95,7 +96,6 @@ import { LearningContent } from './LearningContent.js';
 import { AreaForAdmin, CompetenceForAdmin, ThematicForAdmin, TubeForAdmin } from './LearningContentForAdmin.js';
 import { Membership } from './Membership.js';
 import { NeutralizationAttempt } from './NeutralizationAttempt.js';
-import { OrganizationInvitation } from './OrganizationInvitation.js';
 import { OrganizationInvitedUser } from './OrganizationInvitedUser.js';
 import { OrganizationLearner } from './OrganizationLearner.js';
 import { OrganizationMemberIdentity } from './OrganizationMemberIdentity.js';
