+++ api/tests/acceptance/application/organizations/cancel-organization-invitation-delete_test.js
@@ -1,4 +1,4 @@
-import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
+import { OrganizationInvitation } from '../../../../src/team/domain/models/OrganizationInvitation.js';
 import {
   createServer,
   databaseBuilder,
