+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -1,9 +1,9 @@
 import lodash from 'lodash';
 
 import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
 import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
 import { ORGANIZATION_FEATURE } from '../../../../src/shared/domain/constants.js';
+import { OrganizationInvitation } from '../../../../src/team/domain/models/OrganizationInvitation.js';
 import {
   createServer,
   databaseBuilder,
